+++ core/server/api_container/main.go
@@ -11,13 +11,11 @@ import (
 	"github.com/kurtosis-tech/kurtosis/api/golang/core/kurtosis_core_rpc_api_bindings"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/backend_creator"
-	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend/shared_helpers"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/enclave"
 	"github.com/kurtosis-tech/kurtosis/core/launcher/args"
 	"github.com/kurtosis-tech/kurtosis/core/launcher/args/kurtosis_backend_config"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server"
-	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/facts_engine"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/module_store"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/module_store/module_launcher"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/service_network"
@@ -182,14 +180,9 @@ func runMain() error {
 		}
 	}()
 
-	db, err := shared_helpers.GetLocalDatabase()
-	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred opening local database")
-	}
-	factsEngine := facts_engine.NewFactsEngine(db, serviceNetwork)
 	// TODO: Consolidate Interpreter, Validator and Executor into a single interface
 	startosisRunner := startosis_engine.NewStartosisRunner(
-		startosis_engine.NewStartosisInterpreterWithFacts(serviceNetwork, factsEngine, gitPackageContentProvider, runtime_value_store.NewRuntimeValueStore()),
+		startosis_engine.NewStartosisInterpreter(serviceNetwork, gitPackageContentProvider, runtime_value_store.NewRuntimeValueStore()),
 		startosis_engine.NewStartosisValidator(&kurtosisBackend, serviceNetwork),
 		startosis_engine.NewStartosisExecutor())
 
@@ -198,7 +191,6 @@ func runMain() error {
 		filesArtifactStore,
 		serviceNetwork,
 		moduleStore,
-		factsEngine,
 		startosisRunner,
 		metricsClient,
 		gitPackageContentProvider,